feat: add action logs to create, delete and update routes

+++ backend/src/routes/user/delete-user.ts
@@ -1,4 +1,5 @@
 import { users } from "@db/schema";
+import { createActionLog } from "@routes/helpers/log";
 import { eq } from "drizzle-orm";
 import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { db } from "index";
@@ -8,11 +9,21 @@ export async function deleteUser(app: FastifyInstance) {
     const { id } = request.user!;
 
     if (!id) {
-      return reply.badRequest('Erro ao deletar usuário!');
+      return reply.badRequest('Erro ao excluir usuário!');
     }
 
-    await db.delete(users).where(eq(users.id, Number(id)));
+    const user = await db.select().from(users).where(eq(users.id, Number(id)));
 
-    return reply.ok('Usuário deletado com sucesso!');
+    if (!user) {
+      return reply.notFound('Usuário não encontrado!');
+    }
+
+    await db.transaction(async (tx) => {
+      await db.delete(users).where(eq(users.id, Number(id)));
+
+      await createActionLog('delete', request, tx, `Usuário de ID ${id} excluiu a conta`);
+    });
+
+    return reply.ok('Usuário excluído com sucesso!');
   })
 }
